intelliapply-api/requirements.txt
fastapi
uvicorn[standard]
arq
redis<6
google-generativeai
python-dotenv
supabase
python-jobspy
pandas
fastapi-cors

# --- New AI Crew Dependencies ---
crewai
crewai[google-genai]
crewai-tools
litellm
langchain-google-genai
langchain-core

# --- New Tika Dependency ---


# --- New Phase 1 Dependencies ---
pdfplumber
python-docx
pymupdf
xhtml2pdf
jinja2